ENH: Adding SEMMacroBuild
